Update to sortphotos.py Date Created correction
A while back, an errant backup program ruined all my creation dates and modified dates on photos.
I sought out this program because it could access the EXIF data to obtain the true image capture time.
Beyond using the program to organize my photos, I also added coding that enabled the program to take the oldest EXIF date it found and use that as the "Access and Modification Date" for the photo on my Mac. My Mac then automatically changed the Date Created to match the earlier date.